+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### Version 0.2.0
+
+- Add two features:
+    - Image wrapper - this is used to load pngs, etc into `Image`s using the [Image](https://github.com/image-rs/image) crate
+    - Window prefs - provides functions to save and restore window positions
+- Add examples
+- Changes `PixelWrapper` to use `isize` instead of `usize` generally
+
 ### Version 0.1.1
 
 - Add draw_circle and draw_circle_filled